+++ classes/forms/class.ilMumieTaskLPSettingsFormGUI.php
@@ -7,9 +7,9 @@
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
- /**
-  * This form is used to edit the Learning Progress settings of MumieTasks
-  */
+/**
+ * This form is used to edit the Learning Progress settings of MumieTasks
+ */
 class ilMumieTaskLPSettingsFormGUI extends ilPropertyFormGUI
 {
     public function __construct($disable_grade_pool_selection)
@@ -35,7 +35,7 @@ class ilMumieTaskLPSettingsFormGUI extends ilPropertyFormGUI
         $this->addItem($this->modus_item);
 
         $this->gradepool_item = new ilRadioGroupInputGUI($lng->txt('rep_robj_xmum_frm_privategradepool'), "privategradepool");
-        
+
         $this->gradepool_item->setInfo($this->getGradepoolInfo());
         $gradepool_option_true = new ilRadioOption($lng->txt('rep_robj_xmum_frm_enable'), 0);
         $gradepool_option_false = new ilRadioOption($lng->txt('rep_robj_xmum_frm_disable'), 1);
@@ -44,7 +44,7 @@ class ilMumieTaskLPSettingsFormGUI extends ilPropertyFormGUI
         $gradepool_option_false->setDisabled($this->disable_grade_pool_selection);
         $this->gradepool_item->addOption($gradepool_option_true);
         $this->gradepool_item->addOption($gradepool_option_false);
-        if(!$this->disable_grade_pool_selection) {
+        if (!$this->disable_grade_pool_selection) {
             $this->gradepool_item->addOption($gradepool_option_pending);
         }
         $this->addItem($this->gradepool_item);
@@ -62,11 +62,9 @@ class ilMumieTaskLPSettingsFormGUI extends ilPropertyFormGUI
     {
         global $lng;
         $gradepool_info = $lng->txt('rep_robj_xmum_frm_privategradepool_desc') . '<br><br>';
-        if (!$this->disable_grade_pool_selection)
-        {
+        if (!$this->disable_grade_pool_selection) {
             $gradepool_info .= $lng->txt('rep_robj_xmum_frm_privategradepool_undecided');
-        } else
-        {
+        } else {
             $gradepool_info .= $lng->txt('rep_robj_xmum_frm_privategradepool_decided');
         }
 
@@ -76,7 +74,7 @@ class ilMumieTaskLPSettingsFormGUI extends ilPropertyFormGUI
 
     public function checkInput()
     {
-        $ok = parent::checkInput();  
+        $ok = parent::checkInput();
         return $ok;
     }
 }