+++ sql/dbupdate.php
@@ -165,7 +165,6 @@ if ($ilDB->numRows($result) < 1) {
 ?>
 <#6>
 <?php
-
 $set = $ilDB->query("SELECT obj_id FROM object_data WHERE type='typ' AND title = 'xmum'");
 if ($rec = $ilDB->fetchAssoc($set)) {
     $typ_id = $rec["obj_id"];
@@ -242,15 +241,17 @@ if (!$ilDB->fetchAssoc($ilDB->query($query))) {
 ?>
 <#9>
 <?php
-if (!$ilDB->tableColumnExists("xmum_mumie_task","privategradepool")) {
+if (!$ilDB->tableColumnExists("xmum_mumie_task", "privategradepool")) {
     $ilDB->addTableColumn(
-        "xmum_mumie_task" , "privategradepool",
+        "xmum_mumie_task",
+        "privategradepool",
         array(
             'type' => 'integer',
             'length' => '2',
             'notnull' => true,
             'default' => '-1'
-    ));
+    )
+    );
 }
 $ilDB->manipulate('UPDATE xmum_mumie_task SET privategradepool = 0');
 ?>
@@ -282,23 +283,49 @@ $fieldsHashes = array(
 $ilDB->createTable("xmum_id_hashes", $fieldsHashes);
 $ilDB->addPrimaryKey("xmum_id_hashes", array("id"));
 $ilDB->createSequence("xmum_id_hashes");
-if (!is_null($result))
-{
-    while ($row = $ilDB->fetchAssoc($result))  {
+if (!is_null($result)) {
+    while ($row = $ilDB->fetchAssoc($result)) {
         $query = 'INSERT INTO xmum_id_hashes (id, usr_id, hash) VALUES (' .
             $ilDB->nextID('xmum_id_hashes') . "," .
             $ilDB->quote($row['usr_id'], 'integer') . "," .
             $ilDB->quote($row['hash'], 'text') . ")";
-            $ilDB->manipulate($query); 
+        $ilDB->manipulate($query);
     }
 }
 ?>
 <#11>
 <?php
-$ilDB->modifyTableColumn("xmum_sso_tokens" , "user",
-array(
-    'type' => 'text',
-    'length' => 143,
-    'notnull' => true,
-));
+$ilDB->modifyTableColumn(
+    "xmum_sso_tokens",
+    "user",
+    array(
+        'type' => 'text',
+        'length' => 143,
+        'notnull' => true,
+    )
+);
+?>
+<#12>
+<?php
+if (!$ilDB->tableExists('xmum_grade_override')) {
+    $gradeOverrideSettings = array(
+        'task_id' => array(
+            'type' => 'integer',
+            'length' => 8,
+            'notnull' => true,
+        ),
+        'usr_id' => array(
+            'type' => 'text',
+            'length' => '143',
+            'notnull' => true,
+        ),
+        'new_grade' => array(
+            'type' => 'integer',
+            'length' => '3',
+            'notnull' => true,
+            'default' => '0'
+        )
+    );
+    $ilDB->createTable("xmum_grade_override", $gradeOverrideSettings);
+}
 ?>