+++ changelog.md
@@ -1,6 +1,11 @@
 # MumieTask - Changelog
 
 All important changes to this plugin will be documented in this file.
+## [v2.3.0] 
+### Added
+- Teachers can now open a user overview page in MUMIE Tasks on which they can see the current grades of all user
+- Teachers can now open a grade overview page with all the submissions and submission dates from a user
+- Teachers can now choose which grade should be used when grading this task
 
 ## [v2.2.1] - 2022-11-18
 ### Changed